requirements.txt
ezodf~=0.3.2
openai~=1.30.5
python-docx~=1.1.2
uvicorn~=0.30.0
fastapi~=0.111.0
